Some reviewer's comments implemented.
