+++ magda/testing/utils.py
@@ -4,7 +4,7 @@ from typing import Callable
 from magda.module import Module
 
 
-def wrap_into_result(result, name='', src_class=None, expose=None):
+def wrap_into_result(result, name='testing-module', src_class=None, expose=None):
     interface = result.__class__
     return Module.Result(
         result=result,
